Mouse support for chess board
The board can now be interacted with using the mouse, in both local and online modes.
Clicking on a piece will reveal possible moves, marked with a dot.
If the opponent's piece can be captured with your currently selected piece, it will blink.
Clicking a selected piece again will toggle its selected status, and clicking a possible move will perform that move.